requirements.txt
yfinance
pandas
numpy   
scikit-learn
requests
textblob
finnhub-python
fredapi